Week 4: Movie app (#3)
* Created Project

* Updated .gitignore

* Set up RecyclerView with CardView
Next is to try and make this a Grid style Layout

* Added GridView and Pairs of Names and URLs
I'm using Picasso to grab the pictures.
Down on the row with 'God's Not Dead' and "TPOBWF" there is a bug there.
I'll have to find a way to fix that later.

* Created DetailActivity
I think I'm going to transfer this over to a fragment later. It'd be good practice

* Back Log?

* Added the wont ACtivity Type
Replaced it with Empty ACtivity

* Updated Code to use Movie and MovieData
I'm not sure if I used the companion object in the best way. I'm going to ask for some advice

* Forgot to Pull Updated README

* Added clickListener and use of a Parcelable
I couldn't remember all the steps to this. So I had to read
a few tutorials.

* Got the DetailActivity to finally work
I made a layout that I'm pretty happy with, and I think looks really nice.
Since the basic code is all done (other than converting it to fragments) I'm going to
remake the main layout to look a bit nicer.

* FAB now takes you to the IMDb page

* Moved Some Hardcoded Values into 'Values'

* Added Another Movie for Testing

* Added 'About Menu'
I wan't to add more, but I think that if I add more information for the movies
that the hardcoded text will be too much. For everythign I want to add I'd probably need
to use an API. I desided agains tusing it because it seemed like jen didn't want us to use
it just yet, incase we expand on this later.

* Added App Icon

* Added Advice from PR