Week5: Additions to Movie App (#4)
* Implemented JetPack and Fragments
I'm still going to do 'clean-up' but I wanted a working commit before I did

* Just Cleanup

* Updated WEEKFIVE.md

* Added Landscape Layout for fragment_detail
It isn't the prettiest but it will do for now.

* Basic Layout for Login Screen

* Login Functionality

* Logout Functionality
I'm pretty sure the way that I did this might be wrong. But I'm not sure how to not do it in the
MainActivity. Plus I'm not sure if how I navigated back to the login screen is ok practice.
I probaly either ask or do some research back to this.

* Added Room Database for Movies
This was pretty difficult to grasp. I think that the lesson videos were poorly made.
When he was trying to teach how to use the room there was so much code that was written for you
that it just made it a lot more confusing. So I found a good lesson at: https://codelabs.developers.google.com/codelabs/android-room-with-a-view-kotlin/#0

I might add a database to remember useres just so I can practice it more.
I'm not sure how to seperete everything in different packages. So if you have advice for that, that would be awesome.

* Updated CheckList

* Added Register and Horz Views

* Added Logic For Register

* Fixed Pixel Poster
For some reason the links that I'm using to get the Pixel Poster keep getting taken down?
If the picture doesn't show up that is problem what happend again. It has happend three times now.

* Test Commit

* Updated UserDataManager to make a user Shared Pref

* Acces values not entire user

* Fixed Smaller Issues

* Finished Up Requests