Movie App: Week 6 (#5)
* Started working on new Layouts (not done)
The testFragment is just when I'm trying to display the new_login_fragment. I'm going to focus on some code before
continuing with layouts.

* Moved Around Packages and Added UserDatabase instead of shared prefs
Still trying to figure out how to log out the user.

* Added LogOut and Login Functionality with Room Database

* Implemented Coroutines

* Updated CheckList

* Working Commit
This still has the runBlocking to populate the DetailFragment. I'm starting to run out of time. I know that I can use the LiveData but I don't have time to implement it.

* Revised Code