Week7: Api Assignment (#7)
* Created Week7 ReadMe

* Created Project

* Set up basic grid recycler view and a basic way to read http
I'm extreamly confused on this API stuff. I think I just didn't understand the views that much.
I rewatched them probably 3 times by now and I'm still not getting it. This is going to take a lot of
doc reading and trial and error.

* Now takes first gen pokemon (first 151 pokemon) and displays them in a recycler view
What I want to do is find out how to get the types of the pokemon so then I can change the color of the card
to match the type of the pokemon. I still need to add them to a room database, or atleast add a favorites page?
Not sure yet.

* Now fetches sprites and types of pokemon
I tried to change the background color of the card to match the type but it bugs out for some reason.
I think I'm going to put them in a database and then update the recycler view. That is my first idea
but it might change. Not sure.

* Updated how I'm displaying the Pokemon
It seems like this runs faster than what I had before.

* Api now saves onto a database
When the app first runs, if there is no internet connection it will crash the app.
But if you run the app at first with connection, and then without connection it works fine then.

* App no longer crashes when there isn't internet connection
There isn't an error message. My first try at the error message didn't work.
So I'll have to play with it.

* Updated Logo,Name, and Cleaned Code

* Added About Menu

* Added dynamic card backgrounds.
The backgrounds will change depending on the type of the pokemon.

* Fixed some loading images, and figured out Glide

* Code Clean Up. No changes

* Revised Code