Week8: WorkManagers (#8)
* Moved and Renamed Project from 'Week-7/ApiAssignment/' to 'PokeDex'

* Updated .gitignore

* Moved Project To Its Own Folder

* The App now uses a Work Manager to download the pokemon from the api
It is a tad bit slower, but the way that I populate the recycler view it doesn't actually seem like it is slower
I'm going to leave the in the Logs so then you can see how fast/slow they are actually loading in.
I need to do some reasearch into doing this "refresh every hour" requirement.

Also I looked into changing the api to get a list of pokemon instead off doing 151 network calls. However the way that
it gives you the list of pokemon I would have to do a bunch of calls just to get the pokemon type and sprite anyway.
Here is a link if you want to check it out.

Pokemon Api: https://pokeapi.co/
Call to get first 151 Pokemon: https://pokeapi.co/api/v2/pokemon/?limit=151

* Ability to switch between generations
There is a bug that if the pokemon are loading in and then you switch while it is loading
I'll have to try and figure out how to fix that. I'm not sure if what I'm doing is the best practice,
but I'm going to be chancing the "nukeDatabase" into its own worker since I'm going to need it for the
'refresh'.

* Implemented Periodic Worker
I'm a little confused, because what the periodic worker is doing, is basically the exact same that
the download worker is doing. I'm not sure if this will meet expectations for what is needed.

* Moved around code, and fixed bug with PW

* Added colors for Fairy, Steel, and Dark types

* Just style and color changes

* Added Toast Messages to Downloaders
I couldn't figure out how to use the DownloadManager so I could show a notification. That might be something I
implement later, but the requirments did say that it could either be a notification or a toast.
I still need to fix that bug while downloading and then switching generations. My first attempt was to make the
'currentGen' variable into a LiveData so then I could observe it from the worker, and if it changes I could cancel
current worker and start it up again, but that wasn't wanting to work.

* Changed HardCoded Values to use values

* Just Code Clean-Up

* Added App Icon

* Code Revisions