Week 9: MVVM (#9)
* Added PokemonRepository
The repo now does all of the 'talking' between the Api and the Database

* PokemonViewModel now does most of the work
I was kind of just playing around with the Documentation. I know some of it is pretty redundent
But I wanted to just try it out.

* Added Tests for PokemonViewModel and PokemonRepository
I'm not sure if I'm supposed to add anything else for this weeks requirements.
I've gone over it probably a good 20 times now and I'm just not sure. I'm pretty sure I met all of the requirements
but I'm somewhat nervious that I missed something.